grokking file conversion
