+++ Site/helpers.ts
@@ -26,29 +26,10 @@ export interface ManifestInfo {
   warnings: Array<ManifestGuidance>;
 }
 
-export enum ValidContentType {
-  webmanifest = "application/manifest+json",
-  json = "application/json",
-  binary = "application/octet-stream",
-}
-
 export enum ManifestFormat {
   w3c = "w3c",
   chromeos = "chromeos",
   edgeextension = "edgeextension",
   windows10 = "windows10",
   firefox = "firefox",
-}
-
-export function ifSupportedFile(req: HttpRequest): boolean {
-  switch (req.headers["content-type"]) {
-    case ValidContentType.webmanifest:
-    case ValidContentType.json:
-      return true;
-
-    case ValidContentType.binary:
-    default:
-      // handle binary later
-      return false;
-  }
 }
